feat: Estavel com banco de dados

+++ src/main/java/br/com/edu/utfpr/muralpostit/service/EmployerService.java
@@ -3,7 +3,7 @@ package br.com.edu.utfpr.muralpostit.service;
 import br.com.edu.utfpr.muralpostit.model.dao.EmployerDAO;
 import br.com.edu.utfpr.muralpostit.model.domain.Employer;
 
-public class EmployerService extends AbstractService<Long, Employer>{
+public class EmployerService extends AbstractService<Integer, Employer>{
 
     public EmployerService() {
         dao = new EmployerDAO();
